+++ script/tool/lib/src/xctest_command.dart
@@ -17,8 +17,10 @@ const String _kXCRunCommand = 'xcrun';
 const String _kFoundNoSimulatorsMessage =
     'Cannot find any available simulators, tests failed';
 
-/// The command to run iOS XCTests in plugins, this should work for both XCUnitTest and XCUITest targets.
-/// The tests target have to be added to the xcode project of the example app. Usually at "example/ios/Runner.xcworkspace".
+/// The command to run XCTests (XCUnitTest and XCUITest) in plugins.
+/// The tests target have to be added to the Xcode project of the example app,
+/// usually at "example/{ios,macos}/Runner.xcworkspace".
+///
 /// The static analyzer is also run.
 class XCTestCommand extends PluginCommand {
   /// Creates an instance of the test command.
@@ -33,52 +35,61 @@ class XCTestCommand extends PluginCommand {
           'this is passed to the `-destination` argument in xcodebuild command.\n'
           'See https://developer.apple.com/library/archive/technotes/tn2339/_index.html#//apple_ref/doc/uid/DTS40014588-CH1-UNIT for details on how to specify the destination.',
     );
+    argParser.addFlag(kPlatformFlagIos, help: 'Runs the iOS tests');
+    argParser.addFlag(kPlatformFlagMacos, help: 'Runs the macOS tests');
   }
 
   @override
   final String name = 'xctest';
 
   @override
-  final String description = 'Runs the xctests in the iOS example apps.\n\n'
+  final String description =
+      'Runs the xctests in the iOS and/or macOS example apps.\n\n'
       'This command requires "flutter" and "xcrun" to be in your path.';
 
   @override
   Future<void> run() async {
-    String destination = getStringArg(_kiOSDestination);
-    if (destination.isEmpty) {
-      final String? simulatorId = await _findAvailableIphoneSimulator();
-      if (simulatorId == null) {
-        print(_kFoundNoSimulatorsMessage);
-        throw ToolExit(1);
+    final bool testIos = getBoolArg(kPlatformFlagIos);
+    final bool testMacos = getBoolArg(kPlatformFlagMacos);
+
+    if (!(testIos || testMacos)) {
+      print('At least one platform flag must be provided.');
+      throw ToolExit(2);
+    }
+
+    List<String> iosDestinationFlags = <String>[];
+    if (testIos) {
+      String destination = getStringArg(_kiOSDestination);
+      if (destination.isEmpty) {
+        final String? simulatorId = await _findAvailableIphoneSimulator();
+        if (simulatorId == null) {
+          print(_kFoundNoSimulatorsMessage);
+          throw ToolExit(1);
+        }
+        destination = 'id=$simulatorId';
       }
-      destination = 'id=$simulatorId';
+      iosDestinationFlags = <String>[
+        '-destination',
+        destination,
+      ];
     }
 
     final List<String> failingPackages = <String>[];
     await for (final Directory plugin in getPlugins()) {
-      // Start running for package.
       final String packageName =
           p.relative(plugin.path, from: packagesDir.path);
-      print('Start running for $packageName ...');
-      if (!isIosPlugin(plugin)) {
-        print('iOS is not supported by this plugin.');
-        print('\n\n');
-        continue;
+      print('============================================================');
+      print('Start running for $packageName...');
+      bool passed = true;
+      if (testIos) {
+        passed &= await _testPlugin(plugin, 'iOS',
+            extraXcrunFlags: iosDestinationFlags);
       }
-      for (final Directory example in getExamplesForPlugin(plugin)) {
-        // Running tests and static analyzer.
-        print('Running tests and analyzer for $packageName ...');
-        int exitCode = await _runTests(true, destination, example);
-        // 66 = there is no test target (this fails fast). Try again with just the analyzer.
-        if (exitCode == 66) {
-          print('Tests not found for $packageName, running analyzer only...');
-          exitCode = await _runTests(false, destination, example);
-        }
-        if (exitCode == 0) {
-          print('Successfully ran xctest for $packageName');
-        } else {
-          failingPackages.add(packageName);
-        }
+      if (testMacos) {
+        passed &= await _testPlugin(plugin, 'macOS');
+      }
+      if (!passed) {
+        failingPackages.add(packageName);
       }
     }
 
@@ -95,19 +106,59 @@ class XCTestCommand extends PluginCommand {
     }
   }
 
-  Future<int> _runTests(bool runTests, String destination, Directory example) {
+  /// Runs all applicable tests for [plugin], printing status and returning
+  /// success if the tests passed (or did not exist).
+  Future<bool> _testPlugin(
+    Directory plugin,
+    String platform, {
+    List<String> extraXcrunFlags = const <String>[],
+  }) async {
+    if (!pluginSupportsPlatform(platform.toLowerCase(), plugin,
+        requiredMode: PlatformSupport.inline)) {
+      print('$platform is not implemented by this plugin package.');
+      print('\n');
+      return true;
+    }
+    bool passing = true;
+    for (final Directory example in getExamplesForPlugin(plugin)) {
+      // Running tests and static analyzer.
+      final String examplePath =
+          p.relative(example.path, from: plugin.parent.path);
+      print('Running $platform tests and analyzer for $examplePath...');
+      int exitCode =
+          await _runTests(true, example, platform, extraFlags: extraXcrunFlags);
+      // 66 = there is no test target (this fails fast). Try again with just the analyzer.
+      if (exitCode == 66) {
+        print('Tests not found for $examplePath, running analyzer only...');
+        exitCode = await _runTests(false, example, platform,
+            extraFlags: extraXcrunFlags);
+      }
+      if (exitCode == 0) {
+        print('Successfully ran $platform xctest for $examplePath');
+      } else {
+        passing = false;
+      }
+    }
+    return passing;
+  }
+
+  Future<int> _runTests(
+    bool runTests,
+    Directory example,
+    String platform, {
+    List<String> extraFlags = const <String>[],
+  }) {
     final List<String> xctestArgs = <String>[
       _kXcodeBuildCommand,
       if (runTests) 'test',
       'analyze',
       '-workspace',
-      'ios/Runner.xcworkspace',
+      '${platform.toLowerCase()}/Runner.xcworkspace',
       '-configuration',
       'Debug',
       '-scheme',
       'Runner',
-      '-destination',
-      destination,
+      ...extraFlags,
       'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
     ];
     final String completeTestCommand =