+++ script/tool/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add a --build-id flag to `firebase-test-lab` instead of hard-coding the use of
   `CIRRUS_BUILD_ID`. `CIRRUS_BUILD_ID` is the default value for that flag, for backward
   compatibility.
+- `xctest` now supports running macOS tests in addition to iOS
+  - **Breaking change**: it now requires an `--ios` and/or `--macos` flag.
 
 ## 0.2.0